requirements.txt
Django==1.10.7
djangorestframework==3.6.3
mysqlclient==1.3.10
pkg-resources==0.0.0
